requirements.txt
flask==3.1.2
flask-cors==6.0.1
pandas==2.3.3
openpyxl==3.1.5
openai==2.6.1
python-dotenv==1.0.0
gunicorn==22.0.0